+++ src/main/java/by/itacademy/project/chumakou/taf/sledopitby/ui/data/UserData.java
@@ -1,10 +1,10 @@
-package by.itacademy.project.chumakou.taf.sledopitby.ui.users;
+package by.itacademy.project.chumakou.taf.sledopitby.ui.data;
 
 public class UserData {
 
     public static final String NAME_VALID = "Ivan";
-    public static final String EMAIL_VALID = "";
-    public static final String PASSWORD_VALID = "";
+    public static final String EMAIL_VALID = "cs8qfcns@freepost.cc";
+    public static final String PASSWORD_VALID = "iyufmcl558";
 
     public static final String EMAIL_INVALIDE = "ss@mail.ru";
     public static final String PASSWORD_INVALIDE = "dd";
