+++ src/main/java/by/itacademy/project/chumakou/taf/sledopitby/ui/steps/CartSteps.java
@@ -3,8 +3,8 @@ package by.itacademy.project.chumakou.taf.sledopitby.ui.steps;
 import by.itacademy.project.chumakou.taf.sledopitby.ui.pages.HomePage;
 import by.itacademy.project.chumakou.taf.sledopitby.ui.pages.LoginPage;
 import by.itacademy.project.chumakou.taf.sledopitby.ui.pages.SubcategoryProductPage;
-import by.itacademy.project.chumakou.taf.sledopitby.ui.users.TestData;
-import by.itacademy.project.chumakou.taf.sledopitby.ui.users.UserData;
+import by.itacademy.project.chumakou.taf.sledopitby.ui.data.TestData;
+import by.itacademy.project.chumakou.taf.sledopitby.ui.data.UserData;
 import org.openqa.selenium.WebDriver;
 
 public class CartSteps extends BaseStep {
