+++ src/test/java/by/itacademy/project/chumakou/taf/sledopitby/ui/tests/HomeTest.java
@@ -7,7 +7,6 @@ public class HomeTest extends BaseTest{
 
     @Test()
     @DisplayName("Open home page")
-    @Tag("smoke")
     public void testSledopitByOpen() {
         Assertions.assertEquals(HomePage.TITLE_MAIN_PAGE, driver.getTitle());
     }
